moved commands out of main index
